Break apart the release function from the deployment to central function
This is needed for us to further automate the entire build process.  For now, we will run the new release workflow manually and then that will trigger the deployment workflow (once the release plugin creates the tag).

We still need to register a GPG key and setup all the secrets (probably as organization secrets).

The pom has changes for the release plugin so that the profile that does all the work to push to Nexus is no longer done.  There are also a couple of tweaks to the javadoc, gpg and nexus plugins that matches what Dropwizard does and which makes sense (to me at least) for automation.